requirements.txt
absl-py==0.9.0
aioredis==1.2.0
alabaster==0.7.12
aniso8601==7.0.0
apipkg==1.5
apyori==1.1.2
asgiref==3.2.2
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
Babel==2.8.0
bleach==3.1.1
cachetools==4.1.0
certifi==2019.9.11
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.1
django-filter==2.2.0
django-graphql-auth==0.3.10
django-graphql-jwt==0.3.0
django-import-export==2.2.0
django-model-utils==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.16
drf-yasg==1.17.0
et-xmlfile==1.0.1
execnet==1.7.1
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphene==2.1.8
graphene-django==2.9.1
graphene-django-optimizer==0.6.2
graphene-django-plus==2.0
graphene-file-upload==1.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
grpcio==1.28.1
gunicorn==20.0.0
h5py==2.10.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==17.5.0
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
line-bot-sdk==1.15.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
more-itertools==8.2.0
msgpack==0.6.2
mysql-connector-python==8.0.19
numpy==1.18.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
opt-einsum==3.2.1
packaging==19.2
pandas==1.0.3
parso==0.6.2
pep8==1.7.1
Pillow==6.2.1
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.13.1
promise==2.3
protobuf==3.11.3
psycopg2-binary==2.8.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.4
pypandoc==1.4
pyparsing==2.4.2
pytest==5.3.5
pytest-cache==1.0
pytest-cov==2.8.1
pytest-pep8==1.0.6
python-dateutil==2.8.1
python-dotenv==0.11.0
python-http-client==3.2.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python3-openid==3.1.0
pytz==2019.2
pywin32-ctypes==0.2.0
PyYAML==5.2
readme-renderer==24.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
responses==0.10.9
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
Rx==1.6.1
scipy==1.4.1
selenium==3.141.0
sendgrid==3.6.5
sendgrid-django==4.2.0
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.4.3
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tablib==2.0.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
toml==0.10.0
tox==3.14.5
tqdm==4.43.0
twilio==6.36.0
twine==3.1.1
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.25.5
virtualenv==16.7.9
virtualenv-clone==0.5.3
waitress==1.4.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0
zope.interface==4.7.2
